+++ flexx/ui/widgets/_markdown.py
@@ -1,6 +1,6 @@
 """ Markdown widget
 
-Widget containing a string which content gets rendered and shown as markdown text. 
+Widget containing a string which content gets rendered and shown as markdown text.
 
 See the working example from `flexxamples/ui_usage/markdown.py`.
 
@@ -10,7 +10,8 @@ Simple usage:
 
     def init(self):
         content = "# Welcome\n\n" \
-            "Hello.  Welcome to my **website**. This is an example of a widget container for markdown content. " \
+            "Hello.  Welcome to my **website**." \
+            "This is an example of a widget container for markdown content. " \
             "The content can be text or a link.\n\n"
         ui.Markdown(content=content, style='background:#EAECFF;height:60%;')
 
@@ -19,7 +20,9 @@ Simple usage:
 from ... import app, event
 from . import Widget
 
-app.assets.associate_asset(__name__, 'https://cdnjs.cloudflare.com/ajax/libs/showdown/1.9.1/showdown.min.js')
+app.assets.associate_asset(
+    __name__, 'https://cdnjs.cloudflare.com/ajax/libs/showdown/1.9.1/showdown.min.js'
+)
 
 
 class Markdown(Widget):
@@ -40,5 +43,5 @@ class Markdown(Widget):
     @event.reaction
     def __content_change(self):
         global showdown
-        conv = showdown.Converter();
-        self.node.innerHTML = conv.makeHtml(self.content);
+        conv = showdown.Converter()
+        self.node.innerHTML = conv.makeHtml(self.content)