+++ flexxamples/ui_usage/dropdown_container.py
@@ -2,10 +2,10 @@
 Simple use of a dropdown containing a tree widget
 """
 
-from flexx import app, event, ui, flx
+from flexx import flx
 
 
-class Example(ui.Widget):
+class Example(flx.Widget):
 
     CSS = '''
         .flx-DropdownContainer > .flx-TreeWidget {
@@ -15,10 +15,10 @@ class Example(ui.Widget):
 
     def init(self):
         # A nice and cosy tree view
-        with ui.DropdownContainer(text='Scene graph'):
-            with ui.TreeWidget(max_selected=1):
+        with flx.DropdownContainer(text='Scene graph'):
+            with flx.TreeWidget(max_selected=1):
                 for i in range(20):
-                    ui.TreeItem(text='foo %i' % i, checked=False)
+                    flx.TreeItem(text='foo %i' % i, checked=False)
 
 
 if __name__ == '__main__':