+++ flexxamples/ui_usage/label.py
@@ -2,16 +2,16 @@
 Simple use of a label
 """
 
-from flexx import app, event, ui, flx
+from flexx import flx
 
 
-class Example(ui.Widget):
+class Example(flx.Widget):
 
     def init(self):
-        self.label = ui.Label(text="Number:")
-        self.label = ui.Label(html="<b> 45 </b>")
+        self.label = flx.Label(text="Number:")
+        self.label = flx.Label(html="<b> 45 </b>")
 
 
 if __name__ == '__main__':
-     m = flx.launch(Example, 'default-browser')
-     flx.run()
+    m = flx.launch(Example, 'default-browser')
+    flx.run()